pyproject.toml
[project]
name = "blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "markdown2[all]>=2.5.3",
    "pyyaml>=6.0.2",
    "aiohttp>=3.13.3",
    "watchfiles>=1.1.1",
]